requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.8.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Django==4.1.1
django-environ==0.9.0
django-filter==22.1
django-leaflet==0.28.2
djangorestframework==3.13.1
djangorestframework-gis==1.0
Fiona==1.8.21
geopandas==0.11.1
h11==0.13.0
idna==3.4
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyparsing==3.0.9
pyproj==3.4.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
selenium==4.4.3
Shapely==1.8.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wget==3.2
wsproto==1.2.0
